+++ src/dom/doc.rs
@@ -2,9 +2,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use std::cell::{RefCell};
+use std::cell::RefCell;
 use std::fmt;
-use std::rc::{Rc};
+use std::rc::Rc;
 
 use parser::parse_svg;
 use writer;